+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this library will be documented in this file.
 
+## 1.1.0 - 05-01-2025
+
+- Added the default gameboy color palettes to the PalettePreset enum
+- Renamed extract to extractFromFile in ImageExtractor
+- Renamed getImages to extractImages in ImageExtractor
+
 ## 1.0.0 - 31-12-2024
 
 First stable release.